+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support params for multiple custom loaders.
+- Method `DataLoaderWidget.get_prefix()` (default returns `None` = automatic prefix).
+
 ## [2.1.2] - 2025-12-06
 
 ### Fixed